test(components): split overlay test modules
Entire-Checkpoint: 1d0f5b71e6fa
