+++ crates/ars-components/tests/proptest_state_machines/overlay/dialog.rs
@@ -1,0 +1,283 @@
+use super::*;
+
+#[derive(Clone, Debug)]
+enum DialogStep {
+    Send(core_dialog::Event),
+    SetProps(core_dialog::Props),
+}
+
+const DIALOG_EFFECTS: &[core_dialog::Effect] = &[
+    core_dialog::Effect::OpenChange,
+    core_dialog::Effect::FocusInitial,
+    core_dialog::Effect::FocusFirstTabbable,
+    core_dialog::Effect::ScrollLockAcquire,
+    core_dialog::Effect::ScrollLockRelease,
+    core_dialog::Effect::SetBackgroundInert,
+    core_dialog::Effect::RemoveBackgroundInert,
+    core_dialog::Effect::RestoreFocus,
+];
+
+fn arb_focus_target() -> impl Strategy<Value = Option<FocusTarget>> {
+    prop_oneof![
+        Just(None),
+        Just(Some(FocusTarget::First)),
+        Just(Some(FocusTarget::Last)),
+        Just(Some(FocusTarget::AutofocusMarked)),
+        Just(Some(FocusTarget::PreviouslyActive)),
+    ]
+}
+
+fn arb_dialog_role() -> impl Strategy<Value = core_dialog::Role> {
+    prop_oneof![
+        Just(core_dialog::Role::Dialog),
+        Just(core_dialog::Role::AlertDialog)
+    ]
+}
+
+fn arb_dialog_props() -> impl Strategy<Value = core_dialog::Props> {
+    (
+        (
+            prop::option::of(any::<bool>()),
+            any::<bool>(),
+            any::<bool>(),
+            any::<bool>(),
+            any::<bool>(),
+            any::<bool>(),
+            any::<bool>(),
+            arb_focus_target(),
+        ),
+        (
+            arb_focus_target(),
+            arb_dialog_role(),
+            0_u8..=10,
+            any::<bool>(),
+            any::<bool>(),
+        ),
+    )
+        .prop_map(
+            |(
+                (
+                    open,
+                    default_open,
+                    modal,
+                    close_on_backdrop,
+                    close_on_escape,
+                    prevent_scroll,
+                    restore_focus,
+                    initial_focus,
+                ),
+                (final_focus, role, title_level, lazy_mount, unmount_on_exit),
+            )| {
+                core_dialog::Props::new()
+                    .id("dialog")
+                    .open(open)
+                    .default_open(default_open)
+                    .modal(modal)
+                    .close_on_backdrop(close_on_backdrop)
+                    .close_on_escape(close_on_escape)
+                    .prevent_scroll(prevent_scroll)
+                    .restore_focus(restore_focus)
+                    .initial_focus(initial_focus)
+                    .final_focus(final_focus)
+                    .role(role)
+                    .title_level(title_level)
+                    .lazy_mount(lazy_mount)
+                    .unmount_on_exit(unmount_on_exit)
+            },
+        )
+}
+
+fn arb_dialog_event() -> impl Strategy<Value = core_dialog::Event> {
+    prop_oneof![
+        Just(core_dialog::Event::Open),
+        Just(core_dialog::Event::Close),
+        Just(core_dialog::Event::Toggle),
+        Just(core_dialog::Event::CloseOnBackdropClick),
+        Just(core_dialog::Event::CloseOnEscape),
+        Just(core_dialog::Event::RegisterTitle),
+        Just(core_dialog::Event::RegisterDescription),
+        Just(core_dialog::Event::SyncProps),
+    ]
+}
+
+fn arb_dialog_step() -> impl Strategy<Value = DialogStep> {
+    prop_oneof![
+        arb_dialog_event().prop_map(DialogStep::Send),
+        arb_dialog_props().prop_map(DialogStep::SetProps),
+    ]
+}
+
+fn assert_dialog_state_context_invariants(
+    service: &Service<core_dialog::Machine>,
+) -> TestCaseResult {
+    // 1. State ⇔ context.open invariant.
+    prop_assert_eq!(
+        service.context().open,
+        matches!(service.state(), core_dialog::State::Open)
+    );
+
+    // 2. ID stability — derived from props.id at init and never mutated.
+    prop_assert_eq!(service.context().ids.part("trigger"), "dialog-trigger");
+    prop_assert_eq!(service.context().ids.part("content"), "dialog-content");
+    prop_assert_eq!(service.context().ids.part("title"), "dialog-title");
+    prop_assert_eq!(
+        service.context().ids.part("description"),
+        "dialog-description"
+    );
+
+    Ok(())
+}
+
+fn assert_dialog_send_result_invariants(
+    event: core_dialog::Event,
+    result: &SendResult<core_dialog::Machine>,
+    before_context: &core_dialog::Context,
+) -> TestCaseResult {
+    // 3. Effect-name allow-list — every emitted name is one of the
+    // documented `EFFECT_*` constants.
+    for effect in &result.pending_effects {
+        prop_assert!(
+            DIALOG_EFFECTS.contains(&effect.name),
+            "unexpected effect name: {:?}",
+            effect.name
+        );
+
+        // 4. Payload-free invariant — no metadata leaks through any
+        // emitted intent (the agnostic-core contract).
+        prop_assert!(effect.metadata.is_none());
+    }
+
+    // 5. SyncProps is state-preserving: state never changes.
+    if matches!(event, core_dialog::Event::SyncProps) {
+        prop_assert!(!result.state_changed);
+        prop_assert!(result.pending_effects.is_empty());
+    }
+
+    // 6. Guards: CloseOnBackdropClick / CloseOnEscape MUST NOT change
+    // state when the corresponding `ctx.close_on_*` flag is false.
+    if matches!(event, core_dialog::Event::CloseOnBackdropClick)
+        && !before_context.close_on_backdrop
+    {
+        prop_assert_eq!(result.state_changed, false);
+    }
+    if matches!(event, core_dialog::Event::CloseOnEscape) && !before_context.close_on_escape {
+        prop_assert_eq!(result.state_changed, false);
+    }
+
+    // 7. Register{Title,Description} monotonicity: once `has_title` is
+    // true, it stays true; same for `has_description`. (Catch-all
+    // route in the state machine prevents flipping back.)
+    if matches!(
+        event,
+        core_dialog::Event::RegisterTitle | core_dialog::Event::RegisterDescription
+    ) {
+        prop_assert!(!result.state_changed);
+    }
+
+    // 8. State-flipping events emit `EFFECT_OPEN_CHANGE` exactly when
+    // the state actually changed. Conversely, no-op transitions
+    // (e.g., Open while already Open) do not emit it.
+    let emitted_open_change = result
+        .pending_effects
+        .iter()
+        .any(|e| e.name == core_dialog::Effect::OpenChange);
+
+    let state_actually_flipped = result.state_changed;
+
+    if matches!(
+        event,
+        core_dialog::Event::Open
+            | core_dialog::Event::Close
+            | core_dialog::Event::Toggle
+            | core_dialog::Event::CloseOnBackdropClick
+            | core_dialog::Event::CloseOnEscape
+    ) {
+        prop_assert_eq!(emitted_open_change, state_actually_flipped);
+    } else {
+        prop_assert!(!emitted_open_change);
+    }
+
+    Ok(())
+}
+
+proptest! {
+    #![proptest_config(crate::common::proptest_config())]
+
+    #[test]
+    #[ignore = "proptest — nightly extended-proptest job"]
+    fn proptest_dialog_state_context_invariants_hold(
+        props in arb_dialog_props(),
+        steps in prop::collection::vec(arb_dialog_step(), 0..128),
+    ) {
+        let mut service = Service::<core_dialog::Machine>::new(
+            props,
+            &Env::default(),
+            &core_dialog::Messages::default(),
+        );
+
+        assert_dialog_state_context_invariants(&service)?;
+
+        // Track monotonic flags — once true, must stay true.
+        let mut had_title = service.context().has_title;
+
+        let mut had_description = service.context().has_description;
+
+        for step in steps {
+            match step {
+                DialogStep::Send(event) => {
+                    let before_context = service.context().clone();
+
+                    let result = service.send(event);
+
+                    assert_dialog_send_result_invariants(event, &result, &before_context)?;
+                }
+
+                DialogStep::SetProps(props) => {
+                    let before_open_prop = service.props().open;
+                    let next_open_prop = props.open;
+
+                    drop(service.set_props(props));
+
+                    // After set_props, every context-backed prop in the
+                    // service.props() snapshot must equal the same field
+                    // in `service.context()` — proves SyncProps replayed
+                    // the incoming props.
+                    let p = service.props();
+                    let c = service.context();
+
+                    prop_assert_eq!(c.modal, p.modal);
+                    prop_assert_eq!(c.close_on_backdrop, p.close_on_backdrop);
+                    prop_assert_eq!(c.close_on_escape, p.close_on_escape);
+                    prop_assert_eq!(c.prevent_scroll, p.prevent_scroll);
+                    prop_assert_eq!(c.restore_focus, p.restore_focus);
+                    prop_assert_eq!(c.initial_focus, p.initial_focus);
+                    prop_assert_eq!(c.final_focus, p.final_focus);
+                    prop_assert_eq!(c.role, p.role);
+
+                    // Controlled-open sync: when open prop flipped to a
+                    // concrete value, state and ctx.open must match.
+                    if before_open_prop != next_open_prop
+                        && let Some(open) = service.props().open
+                    {
+                        prop_assert_eq!(service.context().open, open);
+                    }
+                }
+            }
+
+            // Monotonicity invariants enforced after each step.
+            if had_title {
+                prop_assert!(service.context().has_title);
+            }
+
+            if had_description {
+                prop_assert!(service.context().has_description);
+            }
+
+            had_title |= service.context().has_title;
+
+            had_description |= service.context().has_description;
+
+            assert_dialog_state_context_invariants(&service)?;
+        }
+    }
+}
