+++ crates/ars-components/tests/proptest_state_machines/overlay/presence.rs
@@ -1,0 +1,66 @@
+use super::*;
+
+fn arb_presence_props() -> impl Strategy<Value = core_presence::Props> {
+    (any::<bool>(), any::<bool>(), any::<bool>(), any::<bool>()).prop_map(
+        |(present, lazy_mount, skip_animation, reduce_motion)| core_presence::Props {
+            id: "presence".to_string(),
+            present,
+            lazy_mount,
+            skip_animation,
+            reduce_motion,
+        },
+    )
+}
+
+fn arb_presence_event() -> impl Strategy<Value = core_presence::Event> {
+    prop_oneof![
+        Just(core_presence::Event::Mount),
+        Just(core_presence::Event::Unmount),
+        Just(core_presence::Event::ContentReady),
+        Just(core_presence::Event::AnimationEnd),
+    ]
+}
+
+proptest! {
+    #![proptest_config(crate::common::proptest_config())]
+
+    #[test]
+    #[ignore = "proptest — nightly extended-proptest job"]
+    fn proptest_presence_state_context_invariants_hold(
+        props in arb_presence_props(),
+        events in prop::collection::vec(arb_presence_event(), 0..128),
+    ) {
+        let mut service = Service::<core_presence::Machine>::new(props, &Env::default(), &core_presence::Messages);
+
+        for event in events {
+            drop(service.send(event));
+
+            match service.state() {
+                core_presence::State::Unmounted => {
+                    prop_assert!(!service.context().present);
+                    prop_assert!(!service.context().mounted);
+                    prop_assert!(!service.context().unmounting);
+                }
+
+                core_presence::State::Mounting => {
+                    prop_assert!(service.context().present);
+                    prop_assert!(service.context().mounted);
+                    prop_assert!(!service.context().unmounting);
+                }
+
+                core_presence::State::Mounted => {
+                    prop_assert!(service.context().present);
+                    prop_assert!(service.context().mounted);
+                    prop_assert!(!service.context().unmounting);
+                }
+
+                core_presence::State::UnmountPending => {
+                    prop_assert!(!service.context().present);
+                    prop_assert!(service.context().mounted);
+                    prop_assert!(service.context().unmounting);
+                }
+            }
+        }
+    }
+
+}
