+++ crates/ars-components/tests/spec_conformance/overlay/floating_panel.rs
@@ -1,0 +1,55 @@
+use ars_components::overlay::floating_panel;
+use ars_core::{ComponentPart, Env, HtmlAttr, Service};
+
+use crate::helper::assert_anatomy;
+
+#[test]
+fn floating_panel_anatomy_matches_spec() {
+    assert_anatomy(
+        "floating-panel",
+        &[
+            (floating_panel::Part::Root, "root"),
+            (floating_panel::Part::Header, "header"),
+            (floating_panel::Part::DragHandle, "drag-handle"),
+            (floating_panel::Part::Title, "title"),
+            (floating_panel::Part::Content, "content"),
+            (floating_panel::Part::Footer, "footer"),
+            (
+                floating_panel::Part::ResizeHandle {
+                    handle: floating_panel::ResizeHandle::N,
+                },
+                "resize-handle",
+            ),
+            (floating_panel::Part::CloseTrigger, "close-trigger"),
+            (floating_panel::Part::MinimizeTrigger, "minimize-trigger"),
+            (floating_panel::Part::MaximizeTrigger, "maximize-trigger"),
+            (floating_panel::Part::StageTrigger, "stage-trigger"),
+        ],
+    );
+}
+
+#[test]
+fn floating_panel_all_resize_handles_emit_handle_attrs() {
+    let service = Service::<floating_panel::Machine>::new(
+        floating_panel::Props {
+            id: "floating-panel".to_string(),
+            ..floating_panel::Props::default()
+        },
+        &Env::default(),
+        &floating_panel::Messages::default(),
+    );
+
+    let api = service.connect(&|_| {});
+
+    for handle in floating_panel::ResizeHandle::ALL {
+        let part = floating_panel::Part::ResizeHandle { handle };
+
+        let attrs = api.resize_handle_attrs(handle);
+
+        assert_eq!(part.name(), "resize-handle");
+        assert_eq!(
+            attrs.get(&HtmlAttr::Data("ars-handle")),
+            Some(handle.as_str())
+        );
+    }
+}
