+++ crates/ars-components/tests/spec_conformance/overlay/hover_card.rs
@@ -1,0 +1,19 @@
+use ars_components::overlay::hover_card;
+
+use crate::helper::assert_anatomy;
+
+#[test]
+fn hover_card_anatomy_matches_spec() {
+    assert_anatomy(
+        "hover-card",
+        &[
+            (hover_card::Part::Root, "root"),
+            (hover_card::Part::Trigger, "trigger"),
+            (hover_card::Part::Positioner, "positioner"),
+            (hover_card::Part::Content, "content"),
+            (hover_card::Part::Arrow, "arrow"),
+            (hover_card::Part::Title, "title"),
+            (hover_card::Part::DismissButton, "dismiss-button"),
+        ],
+    );
+}
