+++ crates/ars-components/tests/spec_conformance/overlay/mod.rs
@@ -1,0 +1,12 @@
+//! Spec-conformance tests for `crates/ars-components/src/overlay/*`.
+
+mod alert_dialog;
+mod dialog;
+mod drawer;
+mod floating_panel;
+mod hover_card;
+mod popover;
+mod presence;
+mod toast;
+mod tooltip;
+mod tour;
